backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic==0.26.0
sentence-transformers==2.7.0
chromadb==0.5.0
networkx==3.3
PyMuPDF==1.24.3
python-multipart==0.0.9
pydantic==2.7.0
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.4.2
